@@ -41,7 +41,7 @@
     "react": "^15.0.0 || ^16.0.0"
   },
   "dependencies": {
-    "history": "^4.6.3",
+    "history": "^4.7.2",
     "query-string": "^4.3.4",
     "re-reselect": "^0.5.0",
     "reselect": "^3.0.1",
@@ -64,15 +64,15 @@
     "babel-preset-react": "^6.24.1",
     "cross-env": "^5.0.1",
     "css-loader": "^0.28.4",
-    "enzyme": "^3.2.0",
-    "enzyme-adapter-react-16": "^1.1.0",
+    "enzyme": "^3.7.0",
+    "enzyme-adapter-react-16": "^1.6.0",
     "eslint": "^3.19.0",
     "eslint-config-airbnb": "^15.0.2",
     "eslint-plugin-flowtype": "^2.35.0",
     "eslint-plugin-import": "^2.7.0",
     "eslint-plugin-jsx-a11y": "^5.1.1",
     "eslint-plugin-react": "^7.1.0",
-    "flow-bin": "^0.66.0",
+    "flow-bin": "^0.85.0",
     "flow-status-webpack-plugin": "^0.1.7",
     "jasmine-core": "^2.6.4",
     "karma": "^1.7.0",
@@ -85,15 +85,15 @@
     "karma-webpack": "^2.0.4",
     "node-sass": "^4.5.3",
     "prop-types": "^15.6.0",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
-    "react-redux": "^5.0.6",
-    "react-router": "^4.2.0",
-    "react-router-dom": "^4.2.2",
+    "react": "^16.6.0",
+    "react-dom": "^16.6.0",
+    "react-redux": "^5.1.0",
+    "react-router": "^4.3.1",
+    "react-router-dom": "^4.3.1",
     "react-styleguidist": "^6.1.0",
     "react-test-renderer": "^16.2.0",
-    "redux": "^3.7.2",
-    "redux-form": "^7.2.0",
+    "redux": "^4.0.1",
+    "redux-form": "^7.4.2",
     "rimraf": "^2.6.1",
     "semantic-release": "^15.1.2",
     "style-loader": "^0.18.2",
