@@ -4,8 +4,8 @@
 
 import React, { Component } from 'react';
 import type { ComponentType } from 'react';
-import PropTypes from 'prop-types';
 import shallowEqual from 'shallowequal';
+import { QueryManagerContext } from './components/QueryContainer';
 
 const getOptionsFromProps = (options: Object, props: Object) => {
   return Object.keys(options).reduce(
@@ -14,59 +14,65 @@ const getOptionsFromProps = (options: Object, props: Object) => {
 };
 
 const connectQueryToProps =
-  (namespace: string, options: Object, makeRefAvailable: boolean = false) => (InnerComponent: ComponentType<*>) => {
-    return class extends Component<*, *> {
-      static contextTypes = {
-        queryManager: PropTypes.object
-      };
+  (thisNamespace: ?string, options: Object, makeRefAvailable: boolean = false) =>
+    (InnerComponent: ComponentType<*>) => {
+      const namespace = thisNamespace || undefined;
+      class ConnectQueryHoc extends Component<*, *> {
 
-      context: Object;
+        innerComponentRef: ?Element;
 
-      innerComponentRef: ?Element;
-
-      componentWillMount() {
-        const state = this.context.queryManager.register(namespace, options, this.props);
-        this.setState(state);
-      }
-
-      shouldComponentUpdate(nextProps: Object, nextState: Object) {
-        return !shallowEqual(this.state, nextState);
-      }
-
-      componentDidUpdate(prevProps: Object, prevState: Object) {
-        if (this.context.queryManager.isTransitioning()) {
-          return;
+        componentWillMount() {
+          const state = this.props.queryManager.register(namespace, options, this.props);
+          this.setState(state);
         }
-      // check if the parameters actually changed:
-        if (!shallowEqual(
+
+        shouldComponentUpdate(nextProps: Object, nextState: Object) {
+          return !shallowEqual(this.state, nextState);
+        }
+
+        componentDidUpdate(prevProps: Object, prevState: Object) {
+          if (this.props.queryManager.isTransitioning()) {
+            return;
+          }
+        // check if the parameters actually changed:
+          if (!shallowEqual(
           getOptionsFromProps(options, prevState),
           getOptionsFromProps(options, this.state))
       ) {
-          this.context.queryManager.pushChanges(namespace, this.state);
+            this.props.queryManager.pushChanges(namespace, this.state);
+          }
         }
-      }
 
-      componentWillReceiveProps(props: Object) {
-        this.context.queryManager.updateProps(namespace, props);
+      UNSAFE_componentWillReceiveProps(props: Object) { // eslint-disable-line
+        this.props.queryManager.updateProps(namespace, props);
         this.setState(props);
       }
 
-      componentWillUnmount() {
-        if (this.context.queryManager) {
-          this.context.queryManager.unregister(namespace);
+        componentWillUnmount() {
+          if (this.props.queryManager) {
+            this.props.queryManager.unregister(namespace);
+          }
         }
-      }
 
-      getWrappedInstance() {
-        return this.innerComponentRef;
-      }
+        getWrappedInstance() {
+          return this.innerComponentRef;
+        }
 
-      render() {
-        const additionalProps = makeRefAvailable ?
+        render() {
+          const additionalProps = makeRefAvailable ?
           { ref: (instance) => { this.innerComponentRef = instance; } } : {};
-        return <InnerComponent {...additionalProps} {...this.state} />;
-      }
-  };
-  };
+          return <InnerComponent {...additionalProps} {...this.state} />;
+        }
+  }
+      const WithQueryManager = (props: Object) => (
+        <QueryManagerContext.Consumer>
+          {queryManager => (
+            <ConnectQueryHoc queryManager={queryManager} {...props} />
+      )}
+        </QueryManagerContext.Consumer>
+  );
+      return WithQueryManager;
+    };
+
 
 export default connectQueryToProps;
