@@ -1,5 +1,5 @@
-// flow-typed signature: 2344447518e65a3247016f77c83ea024
-// flow-typed version: <<STUB>>/babel-cli_v^6.24.1/flow_v0.61.0
+// flow-typed signature: 0b28d09116249f7c2c30e96f88b8ec42
+// flow-typed version: <<STUB>>/babel-cli_v^6.24.1/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
