@@ -1,5 +1,5 @@
-// flow-typed signature: 58b12a9978527dd32132602d45f30ba4
-// flow-typed version: <<STUB>>/babel-core_v^6.26.0/flow_v0.61.0
+// flow-typed signature: 9ce3e904ad5fec9d3d55c020e090c569
+// flow-typed version: <<STUB>>/babel-core_v^6.26.0/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
