@@ -1,5 +1,5 @@
-// flow-typed signature: 7034b9793ba5ea86706a39ef1bc5c936
-// flow-typed version: <<STUB>>/babel-eslint_v^7.2.3/flow_v0.61.0
+// flow-typed signature: 927b072c44243718f0ee7fc9b1bd68ac
+// flow-typed version: <<STUB>>/babel-eslint_v^7.2.3/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
