@@ -1,5 +1,5 @@
-// flow-typed signature: 9cdda99da32edc6873c88608f0379b75
-// flow-typed version: <<STUB>>/babel-loader_v^7.1.1/flow_v0.61.0
+// flow-typed signature: fc2bc84077646c51b9483259b5a2e787
+// flow-typed version: <<STUB>>/babel-loader_v^7.1.1/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
