@@ -1,5 +1,5 @@
-// flow-typed signature: 336405e328f44ecd1b994627bfeb7dcf
-// flow-typed version: <<STUB>>/babel-plugin-syntax-async-functions_v^6.13.0/flow_v0.61.0
+// flow-typed signature: 497cac1fcc6af53d13057a9c69b399cf
+// flow-typed version: <<STUB>>/babel-plugin-syntax-async-functions_v^6.13.0/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
