@@ -1,5 +1,5 @@
-// flow-typed signature: f44f0d1ae004c824cceaec958e68cacb
-// flow-typed version: <<STUB>>/babel-plugin-syntax-dynamic-import_v^6.18.0/flow_v0.61.0
+// flow-typed signature: fcf3442c9e4ce24e6b910073af49d4a8
+// flow-typed version: <<STUB>>/babel-plugin-syntax-dynamic-import_v^6.18.0/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
