@@ -1,5 +1,5 @@
-// flow-typed signature: 4f9813e3d3380c684c59c9dc9002919b
-// flow-typed version: <<STUB>>/babel-plugin-syntax-export-extensions_v^6.13.0/flow_v0.61.0
+// flow-typed signature: 805932813d00817628353b7476c5b363
+// flow-typed version: <<STUB>>/babel-plugin-syntax-export-extensions_v^6.13.0/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
