@@ -1,5 +1,5 @@
-// flow-typed signature: 979a616e296ac5e5789f35c539460a50
-// flow-typed version: <<STUB>>/babel-plugin-transform-class-properties_v^6.24.1/flow_v0.61.0
+// flow-typed signature: da8fd824ae2982d58ec5252d75e6b2a6
+// flow-typed version: <<STUB>>/babel-plugin-transform-class-properties_v^6.24.1/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
