@@ -1,5 +1,5 @@
-// flow-typed signature: 6d8c8c1e891471727128e762ae449798
-// flow-typed version: <<STUB>>/babel-plugin-transform-flow-strip-types_v^6.22.0/flow_v0.61.0
+// flow-typed signature: f405629f90a407920ccd6fa14aecad63
+// flow-typed version: <<STUB>>/babel-plugin-transform-flow-strip-types_v^6.22.0/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
