@@ -1,5 +1,5 @@
-// flow-typed signature: 128b05565cd4c4fd4e4e210d52a2cd8e
-// flow-typed version: <<STUB>>/babel-plugin-transform-object-rest-spread_v^6.23.0/flow_v0.61.0
+// flow-typed signature: 4a5cff448dc0afd768fddac1147e43ba
+// flow-typed version: <<STUB>>/babel-plugin-transform-object-rest-spread_v^6.23.0/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
