@@ -1,5 +1,5 @@
-// flow-typed signature: 11bec7f3fa7dac6d55e3c722d97391dc
-// flow-typed version: <<STUB>>/babel-plugin-transform-proto-to-assign_v^6.23.0/flow_v0.61.0
+// flow-typed signature: 5c6e8d5a156cd160f4f4b7ae7b5d2fa3
+// flow-typed version: <<STUB>>/babel-plugin-transform-proto-to-assign_v^6.23.0/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
