@@ -1,5 +1,5 @@
-// flow-typed signature: d199216ac1aba5e51f670d9be94d52ac
-// flow-typed version: <<STUB>>/babel-plugin-transform-runtime_v^6.23.0/flow_v0.61.0
+// flow-typed signature: 1d573d82702062a1e2a9ff5a9b71f5cd
+// flow-typed version: <<STUB>>/babel-plugin-transform-runtime_v^6.23.0/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
