@@ -1,5 +1,5 @@
-// flow-typed signature: 49da5d5d4bbd668199a674f7485708e4
-// flow-typed version: <<STUB>>/babel-preset-env_v^1.6.0/flow_v0.61.0
+// flow-typed signature: 947ad691d6951fea2dbcf63a6abe7c44
+// flow-typed version: <<STUB>>/babel-preset-env_v^1.6.0/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
