@@ -1,5 +1,5 @@
-// flow-typed signature: 0579fa410a6e1e10894b84cd1a0fa707
-// flow-typed version: <<STUB>>/babel-preset-react_v^6.24.1/flow_v0.61.0
+// flow-typed signature: 78b177a64b4c84f804fdc1e9b1f6e415
+// flow-typed version: <<STUB>>/babel-preset-react_v^6.24.1/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
