@@ -1,5 +1,5 @@
-// flow-typed signature: 0e0ac480f75ddd77676cbbf707fc9315
-// flow-typed version: <<STUB>>/cross-env_v^5.0.1/flow_v0.61.0
+// flow-typed signature: e62b62cbd824dddf35eca265028863f1
+// flow-typed version: <<STUB>>/cross-env_v^5.0.1/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
