@@ -1,5 +1,5 @@
-// flow-typed signature: 2ae46001ea570fc432434fb68d757b0e
-// flow-typed version: <<STUB>>/css-loader_v^0.28.4/flow_v0.61.0
+// flow-typed signature: e8c1ec9936689797addd0ab25761db7f
+// flow-typed version: <<STUB>>/css-loader_v^0.28.4/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
