@@ -1,5 +1,5 @@
-// flow-typed signature: 5dc9e1e64b141d2bca6d3af7e136ec74
-// flow-typed version: <<STUB>>/enzyme-adapter-react-16_v^1.1.0/flow_v0.61.0
+// flow-typed signature: 55843e1425ba3e0329c31c56ea483f5a
+// flow-typed version: <<STUB>>/enzyme-adapter-react-16_v^1.6.0/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -22,6 +22,10 @@ declare module 'enzyme-adapter-react-16' {
  * require those files directly. Feel free to delete any files that aren't
  * needed.
  */
+declare module 'enzyme-adapter-react-16/build/detectFiberTags' {
+  declare module.exports: any;
+}
+
 declare module 'enzyme-adapter-react-16/build/findCurrentFiberUsingSlowPath' {
   declare module.exports: any;
 }
@@ -31,6 +35,10 @@ declare module 'enzyme-adapter-react-16/build/index' {
 }
 
 declare module 'enzyme-adapter-react-16/build/ReactSixteenAdapter' {
+  declare module.exports: any;
+}
+
+declare module 'enzyme-adapter-react-16/src/detectFiberTags' {
   declare module.exports: any;
 }
 
@@ -47,6 +55,9 @@ declare module 'enzyme-adapter-react-16/src/ReactSixteenAdapter' {
 }
 
 // Filename aliases
+declare module 'enzyme-adapter-react-16/build/detectFiberTags.js' {
+  declare module.exports: $Exports<'enzyme-adapter-react-16/build/detectFiberTags'>;
+}
 declare module 'enzyme-adapter-react-16/build/findCurrentFiberUsingSlowPath.js' {
   declare module.exports: $Exports<'enzyme-adapter-react-16/build/findCurrentFiberUsingSlowPath'>;
 }
@@ -55,6 +66,9 @@ declare module 'enzyme-adapter-react-16/build/index.js' {
 }
 declare module 'enzyme-adapter-react-16/build/ReactSixteenAdapter.js' {
   declare module.exports: $Exports<'enzyme-adapter-react-16/build/ReactSixteenAdapter'>;
+}
+declare module 'enzyme-adapter-react-16/src/detectFiberTags.js' {
+  declare module.exports: $Exports<'enzyme-adapter-react-16/src/detectFiberTags'>;
 }
 declare module 'enzyme-adapter-react-16/src/findCurrentFiberUsingSlowPath.js' {
   declare module.exports: $Exports<'enzyme-adapter-react-16/src/findCurrentFiberUsingSlowPath'>;
