@@ -1,5 +1,5 @@
-// flow-typed signature: 90000b5d9d926ad368329b81d684d910
-// flow-typed version: <<STUB>>/eslint-config-airbnb_v^15.0.2/flow_v0.61.0
+// flow-typed signature: 8fe940e5b12a3ddb76efb5ac005518f6
+// flow-typed version: <<STUB>>/eslint-config-airbnb_v^15.0.2/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
