@@ -1,5 +1,5 @@
-// flow-typed signature: cd118f8263acedd8691699f91b044a0a
-// flow-typed version: <<STUB>>/eslint-plugin-flowtype_v^2.35.0/flow_v0.61.0
+// flow-typed signature: 5b16221d2452d588aa28e764030ffd8c
+// flow-typed version: <<STUB>>/eslint-plugin-flowtype_v^2.35.0/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
