@@ -1,5 +1,5 @@
-// flow-typed signature: ca21af217dd724e25d39be7f48fe171a
-// flow-typed version: <<STUB>>/eslint-plugin-import_v^2.7.0/flow_v0.61.0
+// flow-typed signature: ee48033b11be56605ddd9a7d4e1b215d
+// flow-typed version: <<STUB>>/eslint-plugin-import_v^2.7.0/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
