@@ -1,5 +1,5 @@
-// flow-typed signature: f4005a05a4fdf69d75301ab275b2010f
-// flow-typed version: <<STUB>>/eslint-plugin-jsx-a11y_v^5.1.1/flow_v0.61.0
+// flow-typed signature: 807d31d8a999dfeb1b40339f59e4fd0b
+// flow-typed version: <<STUB>>/eslint-plugin-jsx-a11y_v^5.1.1/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
