@@ -1,5 +1,5 @@
-// flow-typed signature: 83cc32ddecb4e8e75dbe8d88c6d2962a
-// flow-typed version: <<STUB>>/eslint-plugin-react_v^7.1.0/flow_v0.61.0
+// flow-typed signature: baae61afa0ddfb13d3a4489ce1ce5d64
+// flow-typed version: <<STUB>>/eslint-plugin-react_v^7.1.0/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
