@@ -1,5 +1,5 @@
-// flow-typed signature: 24db91f6911e01b8d76db6a0a7594a60
-// flow-typed version: <<STUB>>/eslint_v^3.19.0/flow_v0.61.0
+// flow-typed signature: 44ba91f4e7a1cba2fc45372fc38f1937
+// flow-typed version: <<STUB>>/eslint_v^3.19.0/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
