@@ -1,5 +1,5 @@
-// flow-typed signature: 2fe5cdd639cf8c147cd54dbdbb280c0d
-// flow-typed version: <<STUB>>/flow-status-webpack-plugin_v^0.1.7/flow_v0.61.0
+// flow-typed signature: 7d30591ea00757eccd3504fcb7e8b7e8
+// flow-typed version: <<STUB>>/flow-status-webpack-plugin_v^0.1.7/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
