@@ -1,5 +1,5 @@
-// flow-typed signature: 3681eb39623b22eec7f0e6719c65068b
-// flow-typed version: <<STUB>>/jasmine-core_v^2.6.4/flow_v0.61.0
+// flow-typed signature: ac231bbbf16c8802ac13599ebb547fec
+// flow-typed version: <<STUB>>/jasmine-core_v^2.6.4/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
