@@ -1,5 +1,5 @@
-// flow-typed signature: 8ba0ea5bbb52d0ea12ef4fb8bd44fdb8
-// flow-typed version: <<STUB>>/karma-chrome-launcher_v^2.2.0/flow_v0.61.0
+// flow-typed signature: 7cf334a39621ded26719f42fa43052e8
+// flow-typed version: <<STUB>>/karma-chrome-launcher_v^2.2.0/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
