@@ -1,5 +1,5 @@
-// flow-typed signature: ac01ebd8ec4a7e2ba97a75be76902954
-// flow-typed version: <<STUB>>/karma-coverage_v^1.1.1/flow_v0.61.0
+// flow-typed signature: 1d20c34d521b144bc3f16bcc99f51028
+// flow-typed version: <<STUB>>/karma-coverage_v^1.1.1/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
