@@ -1,5 +1,5 @@
-// flow-typed signature: 921f77e7c11082c883c63f00e2f0ca86
-// flow-typed version: <<STUB>>/karma-jasmine-ajax_v^0.1.13/flow_v0.61.0
+// flow-typed signature: 8d56bdb76ad7b8f99e5d194fcc259657
+// flow-typed version: <<STUB>>/karma-jasmine-ajax_v^0.1.13/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
