@@ -1,5 +1,5 @@
-// flow-typed signature: f972f773444b4b8c5926967ad393a678
-// flow-typed version: <<STUB>>/karma-jasmine_v^1.1.0/flow_v0.61.0
+// flow-typed signature: 5448c566693a57b1989600ab14fe2df9
+// flow-typed version: <<STUB>>/karma-jasmine_v^1.1.0/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
