@@ -1,5 +1,5 @@
-// flow-typed signature: 30b24a32d9f467e3c6922202f7c5cd27
-// flow-typed version: <<STUB>>/karma-junit-reporter_v^1.2.0/flow_v0.61.0
+// flow-typed signature: d027f8d3f035fbc534d9939463bbcf47
+// flow-typed version: <<STUB>>/karma-junit-reporter_v^1.2.0/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
