@@ -1,5 +1,5 @@
-// flow-typed signature: f8ef84140a47c66d80918207f75637ac
-// flow-typed version: <<STUB>>/karma-spec-reporter_v0.0.31/flow_v0.61.0
+// flow-typed signature: cf4b68a3ab3b0b5b42224a57dfc17445
+// flow-typed version: <<STUB>>/karma-spec-reporter_v0.0.31/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
