@@ -1,5 +1,5 @@
-// flow-typed signature: 4f61ca2dfbbad8adf8ff43ad525f1c1d
-// flow-typed version: <<STUB>>/karma-webpack_v^2.0.4/flow_v0.61.0
+// flow-typed signature: fa118a2318546da13b09aadf610a2219
+// flow-typed version: <<STUB>>/karma-webpack_v^2.0.4/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
