@@ -1,5 +1,5 @@
-// flow-typed signature: 9c2396f2027c8fc98c655e12d14b25a9
-// flow-typed version: <<STUB>>/karma_v^1.7.0/flow_v0.61.0
+// flow-typed signature: f34851c382639225315535427936cd18
+// flow-typed version: <<STUB>>/karma_v^1.7.0/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
