@@ -1,5 +1,5 @@
-// flow-typed signature: 4c1d638e5810d29ea3d57e2de498dc29
-// flow-typed version: <<STUB>>/node-sass_v^4.5.3/flow_v0.61.0
+// flow-typed signature: 56dab1fb914514146197d6859d7a8bb6
+// flow-typed version: <<STUB>>/node-sass_v^4.5.3/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
