@@ -1,5 +1,5 @@
-// flow-typed signature: 0f94b7c27a8b8b95ce1d2f688ca69f29
-// flow-typed version: <<STUB>>/query-string_v^4.3.4/flow_v0.61.0
+// flow-typed signature: 1f43db5da0ca437f4bd4d524358d91ab
+// flow-typed version: <<STUB>>/query-string_v^4.3.4/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
