@@ -1,5 +1,5 @@
-// flow-typed signature: 6e741c60fe84f77421d45f1278bd3c07
-// flow-typed version: <<STUB>>/re-reselect_v^0.5.0/flow_v0.61.0
+// flow-typed signature: fcec3e38f30c770b46bdd8dd239ea5d9
+// flow-typed version: <<STUB>>/re-reselect_v^0.5.0/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
