@@ -1,5 +1,5 @@
-// flow-typed signature: 1e6728f0a649edac3689d6e2db7487a7
-// flow-typed version: 01716df816/react-router_v4.x.x/flow_>=v0.53.x
+// flow-typed signature: e15aeed0d3686f71822b54cde7b71c83
+// flow-typed version: fbf3e77efa/react-router_v4.x.x/flow_>=v0.63.x
 
 declare module "react-router" {
   // NOTE: many of these are re-exported by react-router-dom and
@@ -53,7 +53,8 @@ declare module "react-router" {
   declare export type ContextRouter = {|
     history: RouterHistory,
     location: Location,
-    match: Match
+    match: Match,
+    staticContext?: StaticRouterContext
   |};
 
   declare export type GetUserConfirmation = (
@@ -90,23 +91,30 @@ declare module "react-router" {
     when?: boolean
   }> {}
 
-  declare export class Redirect extends React$Component<{
+  declare export class Redirect extends React$Component<{|
     to: string | LocationShape,
-    push?: boolean
-  }> {}
+    push?: boolean,
+    from?: string,
+    exact?: boolean,
+    strict?: boolean
+  |}> {}
 
-  declare export class Route extends React$Component<{
+
+  declare export class Route extends React$Component<{|
     component?: React$ComponentType<*>,
     render?: (router: ContextRouter) => React$Node,
     children?: React$ComponentType<ContextRouter> | React$Node,
     path?: string,
     exact?: boolean,
-    strict?: boolean
-  }> {}
+    strict?: boolean,
+    location?: LocationShape,
+    sensitive?: boolean
+  |}> {}
 
-  declare export class Switch extends React$Component<{
-    children?: React$Node
-  }> {}
+  declare export class Switch extends React$Component<{|
+    children?: React$Node,
+    location?: Location
+  |}> {}
 
   declare export function withRouter<P>(
     Component: React$ComponentType<{| ...ContextRouter, ...P |}>
@@ -118,6 +126,7 @@ declare module "react-router" {
     strict?: boolean,
     sensitive?: boolean
   };
+
   declare export function matchPath(
     pathname: string,
     options?: MatchPathOptions | string
