@@ -1,5 +1,5 @@
-// flow-typed signature: a1109cff4c24fe623bedc45a86c4e9bc
-// flow-typed version: <<STUB>>/react-styleguidist_v^6.1.0/flow_v0.61.0
+// flow-typed signature: 5cd31a26d12616293df773d5a7ee9212
+// flow-typed version: <<STUB>>/react-styleguidist_v^6.1.0/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
