@@ -1,92 +1,98 @@
-// flow-typed signature: 149ad0cf68682fb3286583b17404a4db
-// flow-typed version: da30fe6876/reselect_v3.x.x/flow_>=v0.47.x
+// flow-typed signature: 84ab000391e0f17dd212d57ed0b180f5
+// flow-typed version: 5d8678f464/reselect_v3.x.x/flow_>=v0.47.x
 
-// flow-typed signature: 0199525b667f385f2e61dbeae3215f21
-// flow-typed version: b43dff3e0e/reselect_v3.x.x/flow_>=v0.28.x
+type ExtractReturnType = <Return>((...rest: any[]) => Return) => Return;
 
 declare module "reselect" {
-  declare type Selector<-TState, TProps, TResult> = {
-    (state: TState, props: TProps, ...rest: any[]): TResult
-  };
+  declare type InputSelector<-TState, TProps, TResult> =
+    (state: TState, props: TProps, ...rest: any[]) => TResult
+
+  declare type OutputSelector<-TState, TProps, TResult> =
+    & InputSelector<TState, TProps, TResult>
+    & {
+      recomputations(): number,
+      resetRecomputations(): void,
+      resultFunc(state: TState, props: TProps, ...rest: Array<any>): TResult,
+    };
 
   declare type SelectorCreator = {
     <TState, TProps, TResult, T1>(
-      selector1: Selector<TState, TProps, T1>,
+      selector1: InputSelector<TState, TProps, T1>,
       resultFunc: (arg1: T1) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
     <TState, TProps, TResult, T1>(
-      selectors: [Selector<TState, TProps, T1>],
+      selectors: [InputSelector<TState, TProps, T1>],
       resultFunc: (arg1: T1) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
 
     <TState, TProps, TResult, T1, T2>(
-      selector1: Selector<TState, TProps, T1>,
-      selector2: Selector<TState, TProps, T2>,
+      selector1: InputSelector<TState, TProps, T1>,
+      selector2: InputSelector<TState, TProps, T2>,
       resultFunc: (arg1: T1, arg2: T2) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
     <TState, TProps, TResult, T1, T2>(
-      selectors: [Selector<TState, TProps, T1>, Selector<TState, TProps, T2>],
+      selectors: [InputSelector<TState, TProps, T1>, InputSelector<TState, TProps, T2>],
       resultFunc: (arg1: T1, arg2: T2) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
 
     <TState, TProps, TResult, T1, T2, T3>(
-      selector1: Selector<TState, TProps, T1>,
-      selector2: Selector<TState, TProps, T2>,
-      selector3: Selector<TState, TProps, T3>,
+      selector1: InputSelector<TState, TProps, T1>,
+      selector2: InputSelector<TState, TProps, T2>,
+      selector3: InputSelector<TState, TProps, T3>,
       resultFunc: (arg1: T1, arg2: T2, arg3: T3) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
     <TState, TProps, TResult, T1, T2, T3>(
       selectors: [
-        Selector<TState, TProps, T1>,
-        Selector<TState, TProps, T2>,
-        Selector<TState, TProps, T3>
+        InputSelector<TState, TProps, T1>,
+        InputSelector<TState, TProps, T2>,
+        InputSelector<TState, TProps, T3>
       ],
       resultFunc: (arg1: T1, arg2: T2, arg3: T3) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
 
     <TState, TProps, TResult, T1, T2, T3, T4>(
-      selector1: Selector<TState, TProps, T1>,
-      selector2: Selector<TState, TProps, T2>,
-      selector3: Selector<TState, TProps, T3>,
-      selector4: Selector<TState, TProps, T4>,
+      selector1: InputSelector<TState, TProps, T1>,
+      selector2: InputSelector<TState, TProps, T2>,
+      selector3: InputSelector<TState, TProps, T3>,
+      selector4: InputSelector<TState, TProps, T4>,
       resultFunc: (arg1: T1, arg2: T2, arg3: T3, arg4: T4) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
     <TState, TProps, TResult, T1, T2, T3, T4>(
       selectors: [
-        Selector<TState, TProps, T1>,
-        Selector<TState, TProps, T2>,
-        Selector<TState, TProps, T3>,
-        Selector<TState, TProps, T4>
+        InputSelector<TState, TProps, T1>,
+        InputSelector<TState, TProps, T2>,
+        InputSelector<TState, TProps, T3>,
+        InputSelector<TState, TProps, T4>
       ],
       resultFunc: (arg1: T1, arg2: T2, arg3: T3, arg4: T4) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
 
     <TState, TProps, TResult, T1, T2, T3, T4, T5>(
-      selector1: Selector<TState, TProps, T1>,
-      selector2: Selector<TState, TProps, T2>,
-      selector3: Selector<TState, TProps, T3>,
-      selector4: Selector<TState, TProps, T4>,
-      selector5: Selector<TState, TProps, T5>,
+      selector1: InputSelector<TState, TProps, T1>,
+      selector2: InputSelector<TState, TProps, T2>,
+      selector3: InputSelector<TState, TProps, T3>,
+      selector4: InputSelector<TState, TProps, T4>,
+      selector5: InputSelector<TState, TProps, T5>,
       resultFunc: (arg1: T1, arg2: T2, arg3: T3, arg4: T4, arg5: T5) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
     <TState, TProps, TResult, T1, T2, T3, T4, T5>(
       selectors: [
-        Selector<TState, TProps, T1>,
-        Selector<TState, TProps, T2>,
-        Selector<TState, TProps, T3>,
-        Selector<TState, TProps, T4>,
-        Selector<TState, TProps, T5>
+        InputSelector<TState, TProps, T1>,
+        InputSelector<TState, TProps, T2>,
+        InputSelector<TState, TProps, T3>,
+        InputSelector<TState, TProps, T4>,
+        InputSelector<TState, TProps, T5>
       ],
       resultFunc: (arg1: T1, arg2: T2, arg3: T3, arg4: T4, arg5: T5) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
 
     <TState, TProps, TResult, T1, T2, T3, T4, T5, T6>(
-      selector1: Selector<TState, TProps, T1>,
-      selector2: Selector<TState, TProps, T2>,
-      selector3: Selector<TState, TProps, T3>,
-      selector4: Selector<TState, TProps, T4>,
-      selector5: Selector<TState, TProps, T5>,
-      selector6: Selector<TState, TProps, T6>,
+      selector1: InputSelector<TState, TProps, T1>,
+      selector2: InputSelector<TState, TProps, T2>,
+      selector3: InputSelector<TState, TProps, T3>,
+      selector4: InputSelector<TState, TProps, T4>,
+      selector5: InputSelector<TState, TProps, T5>,
+      selector6: InputSelector<TState, TProps, T6>,
       resultFunc: (
         arg1: T1,
         arg2: T2,
@@ -95,15 +101,15 @@ declare module "reselect" {
         arg5: T5,
         arg6: T6
       ) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
     <TState, TProps, TResult, T1, T2, T3, T4, T5, T6>(
       selectors: [
-        Selector<TState, TProps, T1>,
-        Selector<TState, TProps, T2>,
-        Selector<TState, TProps, T3>,
-        Selector<TState, TProps, T4>,
-        Selector<TState, TProps, T5>,
-        Selector<TState, TProps, T6>
+        InputSelector<TState, TProps, T1>,
+        InputSelector<TState, TProps, T2>,
+        InputSelector<TState, TProps, T3>,
+        InputSelector<TState, TProps, T4>,
+        InputSelector<TState, TProps, T5>,
+        InputSelector<TState, TProps, T6>
       ],
       resultFunc: (
         arg1: T1,
@@ -113,16 +119,16 @@ declare module "reselect" {
         arg5: T5,
         arg6: T6
       ) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
 
     <TState, TProps, TResult, T1, T2, T3, T4, T5, T6, T7>(
-      selector1: Selector<TState, TProps, T1>,
-      selector2: Selector<TState, TProps, T2>,
-      selector3: Selector<TState, TProps, T3>,
-      selector4: Selector<TState, TProps, T4>,
-      selector5: Selector<TState, TProps, T5>,
-      selector6: Selector<TState, TProps, T6>,
-      selector7: Selector<TState, TProps, T7>,
+      selector1: InputSelector<TState, TProps, T1>,
+      selector2: InputSelector<TState, TProps, T2>,
+      selector3: InputSelector<TState, TProps, T3>,
+      selector4: InputSelector<TState, TProps, T4>,
+      selector5: InputSelector<TState, TProps, T5>,
+      selector6: InputSelector<TState, TProps, T6>,
+      selector7: InputSelector<TState, TProps, T7>,
       resultFunc: (
         arg1: T1,
         arg2: T2,
@@ -132,16 +138,16 @@ declare module "reselect" {
         arg6: T6,
         arg7: T7
       ) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
     <TState, TProps, TResult, T1, T2, T3, T4, T5, T6, T7>(
       selectors: [
-        Selector<TState, TProps, T1>,
-        Selector<TState, TProps, T2>,
-        Selector<TState, TProps, T3>,
-        Selector<TState, TProps, T4>,
-        Selector<TState, TProps, T5>,
-        Selector<TState, TProps, T6>,
-        Selector<TState, TProps, T7>
+        InputSelector<TState, TProps, T1>,
+        InputSelector<TState, TProps, T2>,
+        InputSelector<TState, TProps, T3>,
+        InputSelector<TState, TProps, T4>,
+        InputSelector<TState, TProps, T5>,
+        InputSelector<TState, TProps, T6>,
+        InputSelector<TState, TProps, T7>
       ],
       resultFunc: (
         arg1: T1,
@@ -152,17 +158,17 @@ declare module "reselect" {
         arg6: T6,
         arg7: T7
       ) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
 
     <TState, TProps, TResult, T1, T2, T3, T4, T5, T6, T7, T8>(
-      selector1: Selector<TState, TProps, T1>,
-      selector2: Selector<TState, TProps, T2>,
-      selector3: Selector<TState, TProps, T3>,
-      selector4: Selector<TState, TProps, T4>,
-      selector5: Selector<TState, TProps, T5>,
-      selector6: Selector<TState, TProps, T6>,
-      selector7: Selector<TState, TProps, T7>,
-      selector8: Selector<TState, TProps, T8>,
+      selector1: InputSelector<TState, TProps, T1>,
+      selector2: InputSelector<TState, TProps, T2>,
+      selector3: InputSelector<TState, TProps, T3>,
+      selector4: InputSelector<TState, TProps, T4>,
+      selector5: InputSelector<TState, TProps, T5>,
+      selector6: InputSelector<TState, TProps, T6>,
+      selector7: InputSelector<TState, TProps, T7>,
+      selector8: InputSelector<TState, TProps, T8>,
       resultFunc: (
         arg1: T1,
         arg2: T2,
@@ -173,17 +179,17 @@ declare module "reselect" {
         arg7: T7,
         arg8: T8
       ) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
     <TState, TProps, TResult, T1, T2, T3, T4, T5, T6, T7, T8>(
       selectors: [
-        Selector<TState, TProps, T1>,
-        Selector<TState, TProps, T2>,
-        Selector<TState, TProps, T3>,
-        Selector<TState, TProps, T4>,
-        Selector<TState, TProps, T5>,
-        Selector<TState, TProps, T6>,
-        Selector<TState, TProps, T7>,
-        Selector<TState, TProps, T8>
+        InputSelector<TState, TProps, T1>,
+        InputSelector<TState, TProps, T2>,
+        InputSelector<TState, TProps, T3>,
+        InputSelector<TState, TProps, T4>,
+        InputSelector<TState, TProps, T5>,
+        InputSelector<TState, TProps, T6>,
+        InputSelector<TState, TProps, T7>,
+        InputSelector<TState, TProps, T8>
       ],
       resultFunc: (
         arg1: T1,
@@ -195,18 +201,18 @@ declare module "reselect" {
         arg7: T7,
         arg8: T8
       ) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
 
     <TState, TProps, TResult, T1, T2, T3, T4, T5, T6, T7, T8, T9>(
-      selector1: Selector<TState, TProps, T1>,
-      selector2: Selector<TState, TProps, T2>,
-      selector3: Selector<TState, TProps, T3>,
-      selector4: Selector<TState, TProps, T4>,
-      selector5: Selector<TState, TProps, T5>,
-      selector6: Selector<TState, TProps, T6>,
-      selector7: Selector<TState, TProps, T7>,
-      selector8: Selector<TState, TProps, T8>,
-      selector9: Selector<TState, TProps, T9>,
+      selector1: InputSelector<TState, TProps, T1>,
+      selector2: InputSelector<TState, TProps, T2>,
+      selector3: InputSelector<TState, TProps, T3>,
+      selector4: InputSelector<TState, TProps, T4>,
+      selector5: InputSelector<TState, TProps, T5>,
+      selector6: InputSelector<TState, TProps, T6>,
+      selector7: InputSelector<TState, TProps, T7>,
+      selector8: InputSelector<TState, TProps, T8>,
+      selector9: InputSelector<TState, TProps, T9>,
       resultFunc: (
         arg1: T1,
         arg2: T2,
@@ -218,18 +224,18 @@ declare module "reselect" {
         arg8: T8,
         arg9: T9
       ) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
     <TState, TProps, TResult, T1, T2, T3, T4, T5, T6, T7, T8, T9>(
       selectors: [
-        Selector<TState, TProps, T1>,
-        Selector<TState, TProps, T2>,
-        Selector<TState, TProps, T3>,
-        Selector<TState, TProps, T4>,
-        Selector<TState, TProps, T5>,
-        Selector<TState, TProps, T6>,
-        Selector<TState, TProps, T7>,
-        Selector<TState, TProps, T8>,
-        Selector<TState, TProps, T9>
+        InputSelector<TState, TProps, T1>,
+        InputSelector<TState, TProps, T2>,
+        InputSelector<TState, TProps, T3>,
+        InputSelector<TState, TProps, T4>,
+        InputSelector<TState, TProps, T5>,
+        InputSelector<TState, TProps, T6>,
+        InputSelector<TState, TProps, T7>,
+        InputSelector<TState, TProps, T8>,
+        InputSelector<TState, TProps, T9>
       ],
       resultFunc: (
         arg1: T1,
@@ -242,19 +248,19 @@ declare module "reselect" {
         arg8: T8,
         arg9: T9
       ) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
 
     <TState, TProps, TResult, T1, T2, T3, T4, T5, T6, T7, T8, T9, T10>(
-      selector1: Selector<TState, TProps, T1>,
-      selector2: Selector<TState, TProps, T2>,
-      selector3: Selector<TState, TProps, T3>,
-      selector4: Selector<TState, TProps, T4>,
-      selector5: Selector<TState, TProps, T5>,
-      selector6: Selector<TState, TProps, T6>,
-      selector7: Selector<TState, TProps, T7>,
-      selector8: Selector<TState, TProps, T8>,
-      selector9: Selector<TState, TProps, T9>,
-      selector10: Selector<TState, TProps, T10>,
+      selector1: InputSelector<TState, TProps, T1>,
+      selector2: InputSelector<TState, TProps, T2>,
+      selector3: InputSelector<TState, TProps, T3>,
+      selector4: InputSelector<TState, TProps, T4>,
+      selector5: InputSelector<TState, TProps, T5>,
+      selector6: InputSelector<TState, TProps, T6>,
+      selector7: InputSelector<TState, TProps, T7>,
+      selector8: InputSelector<TState, TProps, T8>,
+      selector9: InputSelector<TState, TProps, T9>,
+      selector10: InputSelector<TState, TProps, T10>,
       resultFunc: (
         arg1: T1,
         arg2: T2,
@@ -267,19 +273,19 @@ declare module "reselect" {
         arg9: T9,
         arg10: T10
       ) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
     <TState, TProps, TResult, T1, T2, T3, T4, T5, T6, T7, T8, T9, T10>(
       selectors: [
-        Selector<TState, TProps, T1>,
-        Selector<TState, TProps, T2>,
-        Selector<TState, TProps, T3>,
-        Selector<TState, TProps, T4>,
-        Selector<TState, TProps, T5>,
-        Selector<TState, TProps, T6>,
-        Selector<TState, TProps, T7>,
-        Selector<TState, TProps, T8>,
-        Selector<TState, TProps, T9>,
-        Selector<TState, TProps, T10>
+        InputSelector<TState, TProps, T1>,
+        InputSelector<TState, TProps, T2>,
+        InputSelector<TState, TProps, T3>,
+        InputSelector<TState, TProps, T4>,
+        InputSelector<TState, TProps, T5>,
+        InputSelector<TState, TProps, T6>,
+        InputSelector<TState, TProps, T7>,
+        InputSelector<TState, TProps, T8>,
+        InputSelector<TState, TProps, T9>,
+        InputSelector<TState, TProps, T10>
       ],
       resultFunc: (
         arg1: T1,
@@ -293,20 +299,20 @@ declare module "reselect" {
         arg9: T9,
         arg10: T10
       ) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
 
     <TState, TProps, TResult, T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11>(
-      selector1: Selector<TState, TProps, T1>,
-      selector2: Selector<TState, TProps, T2>,
-      selector3: Selector<TState, TProps, T3>,
-      selector4: Selector<TState, TProps, T4>,
-      selector5: Selector<TState, TProps, T5>,
-      selector6: Selector<TState, TProps, T6>,
-      selector7: Selector<TState, TProps, T7>,
-      selector8: Selector<TState, TProps, T8>,
-      selector9: Selector<TState, TProps, T9>,
-      selector10: Selector<TState, TProps, T10>,
-      selector11: Selector<TState, TProps, T11>,
+      selector1: InputSelector<TState, TProps, T1>,
+      selector2: InputSelector<TState, TProps, T2>,
+      selector3: InputSelector<TState, TProps, T3>,
+      selector4: InputSelector<TState, TProps, T4>,
+      selector5: InputSelector<TState, TProps, T5>,
+      selector6: InputSelector<TState, TProps, T6>,
+      selector7: InputSelector<TState, TProps, T7>,
+      selector8: InputSelector<TState, TProps, T8>,
+      selector9: InputSelector<TState, TProps, T9>,
+      selector10: InputSelector<TState, TProps, T10>,
+      selector11: InputSelector<TState, TProps, T11>,
       resultFunc: (
         arg1: T1,
         arg2: T2,
@@ -320,20 +326,20 @@ declare module "reselect" {
         arg10: T10,
         arg11: T11
       ) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
     <TState, TProps, TResult, T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11>(
       selectors: [
-        Selector<TState, TProps, T1>,
-        Selector<TState, TProps, T2>,
-        Selector<TState, TProps, T3>,
-        Selector<TState, TProps, T4>,
-        Selector<TState, TProps, T5>,
-        Selector<TState, TProps, T6>,
-        Selector<TState, TProps, T7>,
-        Selector<TState, TProps, T8>,
-        Selector<TState, TProps, T9>,
-        Selector<TState, TProps, T10>,
-        Selector<TState, TProps, T11>
+        InputSelector<TState, TProps, T1>,
+        InputSelector<TState, TProps, T2>,
+        InputSelector<TState, TProps, T3>,
+        InputSelector<TState, TProps, T4>,
+        InputSelector<TState, TProps, T5>,
+        InputSelector<TState, TProps, T6>,
+        InputSelector<TState, TProps, T7>,
+        InputSelector<TState, TProps, T8>,
+        InputSelector<TState, TProps, T9>,
+        InputSelector<TState, TProps, T10>,
+        InputSelector<TState, TProps, T11>
       ],
       resultFunc: (
         arg1: T1,
@@ -348,7 +354,7 @@ declare module "reselect" {
         arg10: T10,
         arg11: T11
       ) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
 
     <
       TState,
@@ -367,18 +373,18 @@ declare module "reselect" {
       T11,
       T12
     >(
-      selector1: Selector<TState, TProps, T1>,
-      selector2: Selector<TState, TProps, T2>,
-      selector3: Selector<TState, TProps, T3>,
-      selector4: Selector<TState, TProps, T4>,
-      selector5: Selector<TState, TProps, T5>,
-      selector6: Selector<TState, TProps, T6>,
-      selector7: Selector<TState, TProps, T7>,
-      selector8: Selector<TState, TProps, T8>,
-      selector9: Selector<TState, TProps, T9>,
-      selector10: Selector<TState, TProps, T10>,
-      selector11: Selector<TState, TProps, T11>,
-      selector12: Selector<TState, TProps, T12>,
+      selector1: InputSelector<TState, TProps, T1>,
+      selector2: InputSelector<TState, TProps, T2>,
+      selector3: InputSelector<TState, TProps, T3>,
+      selector4: InputSelector<TState, TProps, T4>,
+      selector5: InputSelector<TState, TProps, T5>,
+      selector6: InputSelector<TState, TProps, T6>,
+      selector7: InputSelector<TState, TProps, T7>,
+      selector8: InputSelector<TState, TProps, T8>,
+      selector9: InputSelector<TState, TProps, T9>,
+      selector10: InputSelector<TState, TProps, T10>,
+      selector11: InputSelector<TState, TProps, T11>,
+      selector12: InputSelector<TState, TProps, T12>,
       resultFunc: (
         arg1: T1,
         arg2: T2,
@@ -393,7 +399,7 @@ declare module "reselect" {
         arg11: T11,
         arg12: T12
       ) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
     <
       TState,
       TProps,
@@ -412,18 +418,18 @@ declare module "reselect" {
       T12
     >(
       selectors: [
-        Selector<TState, TProps, T1>,
-        Selector<TState, TProps, T2>,
-        Selector<TState, TProps, T3>,
-        Selector<TState, TProps, T4>,
-        Selector<TState, TProps, T5>,
-        Selector<TState, TProps, T6>,
-        Selector<TState, TProps, T7>,
-        Selector<TState, TProps, T8>,
-        Selector<TState, TProps, T9>,
-        Selector<TState, TProps, T10>,
-        Selector<TState, TProps, T11>,
-        Selector<TState, TProps, T12>
+        InputSelector<TState, TProps, T1>,
+        InputSelector<TState, TProps, T2>,
+        InputSelector<TState, TProps, T3>,
+        InputSelector<TState, TProps, T4>,
+        InputSelector<TState, TProps, T5>,
+        InputSelector<TState, TProps, T6>,
+        InputSelector<TState, TProps, T7>,
+        InputSelector<TState, TProps, T8>,
+        InputSelector<TState, TProps, T9>,
+        InputSelector<TState, TProps, T10>,
+        InputSelector<TState, TProps, T11>,
+        InputSelector<TState, TProps, T12>
       ],
       resultFunc: (
         arg1: T1,
@@ -439,7 +445,7 @@ declare module "reselect" {
         arg11: T11,
         arg12: T12
       ) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
 
     <
       TState,
@@ -459,19 +465,19 @@ declare module "reselect" {
       T12,
       T13
     >(
-      selector1: Selector<TState, TProps, T1>,
-      selector2: Selector<TState, TProps, T2>,
-      selector3: Selector<TState, TProps, T3>,
-      selector4: Selector<TState, TProps, T4>,
-      selector5: Selector<TState, TProps, T5>,
-      selector6: Selector<TState, TProps, T6>,
-      selector7: Selector<TState, TProps, T7>,
-      selector8: Selector<TState, TProps, T8>,
-      selector9: Selector<TState, TProps, T9>,
-      selector10: Selector<TState, TProps, T10>,
-      selector11: Selector<TState, TProps, T11>,
-      selector12: Selector<TState, TProps, T12>,
-      selector13: Selector<TState, TProps, T13>,
+      selector1: InputSelector<TState, TProps, T1>,
+      selector2: InputSelector<TState, TProps, T2>,
+      selector3: InputSelector<TState, TProps, T3>,
+      selector4: InputSelector<TState, TProps, T4>,
+      selector5: InputSelector<TState, TProps, T5>,
+      selector6: InputSelector<TState, TProps, T6>,
+      selector7: InputSelector<TState, TProps, T7>,
+      selector8: InputSelector<TState, TProps, T8>,
+      selector9: InputSelector<TState, TProps, T9>,
+      selector10: InputSelector<TState, TProps, T10>,
+      selector11: InputSelector<TState, TProps, T11>,
+      selector12: InputSelector<TState, TProps, T12>,
+      selector13: InputSelector<TState, TProps, T13>,
       resultFunc: (
         arg1: T1,
         arg2: T2,
@@ -487,7 +493,7 @@ declare module "reselect" {
         arg12: T12,
         arg13: T13
       ) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
     <
       TState,
       TProps,
@@ -507,19 +513,19 @@ declare module "reselect" {
       T13
     >(
       selectors: [
-        Selector<TState, TProps, T1>,
-        Selector<TState, TProps, T2>,
-        Selector<TState, TProps, T3>,
-        Selector<TState, TProps, T4>,
-        Selector<TState, TProps, T5>,
-        Selector<TState, TProps, T6>,
-        Selector<TState, TProps, T7>,
-        Selector<TState, TProps, T8>,
-        Selector<TState, TProps, T9>,
-        Selector<TState, TProps, T10>,
-        Selector<TState, TProps, T11>,
-        Selector<TState, TProps, T12>,
-        Selector<TState, TProps, T13>
+        InputSelector<TState, TProps, T1>,
+        InputSelector<TState, TProps, T2>,
+        InputSelector<TState, TProps, T3>,
+        InputSelector<TState, TProps, T4>,
+        InputSelector<TState, TProps, T5>,
+        InputSelector<TState, TProps, T6>,
+        InputSelector<TState, TProps, T7>,
+        InputSelector<TState, TProps, T8>,
+        InputSelector<TState, TProps, T9>,
+        InputSelector<TState, TProps, T10>,
+        InputSelector<TState, TProps, T11>,
+        InputSelector<TState, TProps, T12>,
+        InputSelector<TState, TProps, T13>
       ],
       resultFunc: (
         arg1: T1,
@@ -536,7 +542,7 @@ declare module "reselect" {
         arg12: T12,
         arg13: T13
       ) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
 
     <
       TState,
@@ -557,20 +563,20 @@ declare module "reselect" {
       T13,
       T14
     >(
-      selector1: Selector<TState, TProps, T1>,
-      selector2: Selector<TState, TProps, T2>,
-      selector3: Selector<TState, TProps, T3>,
-      selector4: Selector<TState, TProps, T4>,
-      selector5: Selector<TState, TProps, T5>,
-      selector6: Selector<TState, TProps, T6>,
-      selector7: Selector<TState, TProps, T7>,
-      selector8: Selector<TState, TProps, T8>,
-      selector9: Selector<TState, TProps, T9>,
-      selector10: Selector<TState, TProps, T10>,
-      selector11: Selector<TState, TProps, T11>,
-      selector12: Selector<TState, TProps, T12>,
-      selector13: Selector<TState, TProps, T13>,
-      selector14: Selector<TState, TProps, T14>,
+      selector1: InputSelector<TState, TProps, T1>,
+      selector2: InputSelector<TState, TProps, T2>,
+      selector3: InputSelector<TState, TProps, T3>,
+      selector4: InputSelector<TState, TProps, T4>,
+      selector5: InputSelector<TState, TProps, T5>,
+      selector6: InputSelector<TState, TProps, T6>,
+      selector7: InputSelector<TState, TProps, T7>,
+      selector8: InputSelector<TState, TProps, T8>,
+      selector9: InputSelector<TState, TProps, T9>,
+      selector10: InputSelector<TState, TProps, T10>,
+      selector11: InputSelector<TState, TProps, T11>,
+      selector12: InputSelector<TState, TProps, T12>,
+      selector13: InputSelector<TState, TProps, T13>,
+      selector14: InputSelector<TState, TProps, T14>,
       resultFunc: (
         arg1: T1,
         arg2: T2,
@@ -587,7 +593,7 @@ declare module "reselect" {
         arg13: T13,
         arg14: T14
       ) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
     <
       TState,
       TProps,
@@ -608,20 +614,20 @@ declare module "reselect" {
       T14
     >(
       selectors: [
-        Selector<TState, TProps, T1>,
-        Selector<TState, TProps, T2>,
-        Selector<TState, TProps, T3>,
-        Selector<TState, TProps, T4>,
-        Selector<TState, TProps, T5>,
-        Selector<TState, TProps, T6>,
-        Selector<TState, TProps, T7>,
-        Selector<TState, TProps, T8>,
-        Selector<TState, TProps, T9>,
-        Selector<TState, TProps, T10>,
-        Selector<TState, TProps, T11>,
-        Selector<TState, TProps, T12>,
-        Selector<TState, TProps, T13>,
-        Selector<TState, TProps, T14>
+        InputSelector<TState, TProps, T1>,
+        InputSelector<TState, TProps, T2>,
+        InputSelector<TState, TProps, T3>,
+        InputSelector<TState, TProps, T4>,
+        InputSelector<TState, TProps, T5>,
+        InputSelector<TState, TProps, T6>,
+        InputSelector<TState, TProps, T7>,
+        InputSelector<TState, TProps, T8>,
+        InputSelector<TState, TProps, T9>,
+        InputSelector<TState, TProps, T10>,
+        InputSelector<TState, TProps, T11>,
+        InputSelector<TState, TProps, T12>,
+        InputSelector<TState, TProps, T13>,
+        InputSelector<TState, TProps, T14>
       ],
       resultFunc: (
         arg1: T1,
@@ -639,7 +645,7 @@ declare module "reselect" {
         arg13: T13,
         arg14: T14
       ) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
 
     <
       TState,
@@ -661,21 +667,21 @@ declare module "reselect" {
       T14,
       T15
     >(
-      selector1: Selector<TState, TProps, T1>,
-      selector2: Selector<TState, TProps, T2>,
-      selector3: Selector<TState, TProps, T3>,
-      selector4: Selector<TState, TProps, T4>,
-      selector5: Selector<TState, TProps, T5>,
-      selector6: Selector<TState, TProps, T6>,
-      selector7: Selector<TState, TProps, T7>,
-      selector8: Selector<TState, TProps, T8>,
-      selector9: Selector<TState, TProps, T9>,
-      selector10: Selector<TState, TProps, T10>,
-      selector11: Selector<TState, TProps, T11>,
-      selector12: Selector<TState, TProps, T12>,
-      selector13: Selector<TState, TProps, T13>,
-      selector14: Selector<TState, TProps, T14>,
-      selector15: Selector<TState, TProps, T15>,
+      selector1: InputSelector<TState, TProps, T1>,
+      selector2: InputSelector<TState, TProps, T2>,
+      selector3: InputSelector<TState, TProps, T3>,
+      selector4: InputSelector<TState, TProps, T4>,
+      selector5: InputSelector<TState, TProps, T5>,
+      selector6: InputSelector<TState, TProps, T6>,
+      selector7: InputSelector<TState, TProps, T7>,
+      selector8: InputSelector<TState, TProps, T8>,
+      selector9: InputSelector<TState, TProps, T9>,
+      selector10: InputSelector<TState, TProps, T10>,
+      selector11: InputSelector<TState, TProps, T11>,
+      selector12: InputSelector<TState, TProps, T12>,
+      selector13: InputSelector<TState, TProps, T13>,
+      selector14: InputSelector<TState, TProps, T14>,
+      selector15: InputSelector<TState, TProps, T15>,
       resultFunc: (
         arg1: T1,
         arg2: T2,
@@ -693,7 +699,7 @@ declare module "reselect" {
         arg14: T14,
         arg15: T15
       ) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
     <
       TState,
       TProps,
@@ -715,21 +721,21 @@ declare module "reselect" {
       T15
     >(
       selectors: [
-        Selector<TState, TProps, T1>,
-        Selector<TState, TProps, T2>,
-        Selector<TState, TProps, T3>,
-        Selector<TState, TProps, T4>,
-        Selector<TState, TProps, T5>,
-        Selector<TState, TProps, T6>,
-        Selector<TState, TProps, T7>,
-        Selector<TState, TProps, T8>,
-        Selector<TState, TProps, T9>,
-        Selector<TState, TProps, T10>,
-        Selector<TState, TProps, T11>,
-        Selector<TState, TProps, T12>,
-        Selector<TState, TProps, T13>,
-        Selector<TState, TProps, T14>,
-        Selector<TState, TProps, T15>
+        InputSelector<TState, TProps, T1>,
+        InputSelector<TState, TProps, T2>,
+        InputSelector<TState, TProps, T3>,
+        InputSelector<TState, TProps, T4>,
+        InputSelector<TState, TProps, T5>,
+        InputSelector<TState, TProps, T6>,
+        InputSelector<TState, TProps, T7>,
+        InputSelector<TState, TProps, T8>,
+        InputSelector<TState, TProps, T9>,
+        InputSelector<TState, TProps, T10>,
+        InputSelector<TState, TProps, T11>,
+        InputSelector<TState, TProps, T12>,
+        InputSelector<TState, TProps, T13>,
+        InputSelector<TState, TProps, T14>,
+        InputSelector<TState, TProps, T15>
       ],
       resultFunc: (
         arg1: T1,
@@ -748,7 +754,7 @@ declare module "reselect" {
         arg14: T14,
         arg15: T15
       ) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
 
     <
       TState,
@@ -771,22 +777,22 @@ declare module "reselect" {
       T15,
       T16
     >(
-      selector1: Selector<TState, TProps, T1>,
-      selector2: Selector<TState, TProps, T2>,
-      selector3: Selector<TState, TProps, T3>,
-      selector4: Selector<TState, TProps, T4>,
-      selector5: Selector<TState, TProps, T5>,
-      selector6: Selector<TState, TProps, T6>,
-      selector7: Selector<TState, TProps, T7>,
-      selector8: Selector<TState, TProps, T8>,
-      selector9: Selector<TState, TProps, T9>,
-      selector10: Selector<TState, TProps, T10>,
-      selector11: Selector<TState, TProps, T11>,
-      selector12: Selector<TState, TProps, T12>,
-      selector13: Selector<TState, TProps, T13>,
-      selector14: Selector<TState, TProps, T14>,
-      selector15: Selector<TState, TProps, T15>,
-      selector16: Selector<TState, TProps, T16>,
+      selector1: InputSelector<TState, TProps, T1>,
+      selector2: InputSelector<TState, TProps, T2>,
+      selector3: InputSelector<TState, TProps, T3>,
+      selector4: InputSelector<TState, TProps, T4>,
+      selector5: InputSelector<TState, TProps, T5>,
+      selector6: InputSelector<TState, TProps, T6>,
+      selector7: InputSelector<TState, TProps, T7>,
+      selector8: InputSelector<TState, TProps, T8>,
+      selector9: InputSelector<TState, TProps, T9>,
+      selector10: InputSelector<TState, TProps, T10>,
+      selector11: InputSelector<TState, TProps, T11>,
+      selector12: InputSelector<TState, TProps, T12>,
+      selector13: InputSelector<TState, TProps, T13>,
+      selector14: InputSelector<TState, TProps, T14>,
+      selector15: InputSelector<TState, TProps, T15>,
+      selector16: InputSelector<TState, TProps, T16>,
       resultFunc: (
         arg1: T1,
         arg2: T2,
@@ -805,7 +811,7 @@ declare module "reselect" {
         arg15: T15,
         arg16: T16
       ) => TResult
-    ): Selector<TState, TProps, TResult>,
+    ): OutputSelector<TState, TProps, TResult>,
     <
       TState,
       TProps,
@@ -828,22 +834,22 @@ declare module "reselect" {
       T16
     >(
       selectors: [
-        Selector<TState, TProps, T1>,
-        Selector<TState, TProps, T2>,
-        Selector<TState, TProps, T3>,
-        Selector<TState, TProps, T4>,
-        Selector<TState, TProps, T5>,
-        Selector<TState, TProps, T6>,
-        Selector<TState, TProps, T7>,
-        Selector<TState, TProps, T8>,
-        Selector<TState, TProps, T9>,
-        Selector<TState, TProps, T10>,
-        Selector<TState, TProps, T11>,
-        Selector<TState, TProps, T12>,
-        Selector<TState, TProps, T13>,
-        Selector<TState, TProps, T14>,
-        Selector<TState, TProps, T15>,
-        Selector<TState, TProps, T16>
+        InputSelector<TState, TProps, T1>,
+        InputSelector<TState, TProps, T2>,
+        InputSelector<TState, TProps, T3>,
+        InputSelector<TState, TProps, T4>,
+        InputSelector<TState, TProps, T5>,
+        InputSelector<TState, TProps, T6>,
+        InputSelector<TState, TProps, T7>,
+        InputSelector<TState, TProps, T8>,
+        InputSelector<TState, TProps, T9>,
+        InputSelector<TState, TProps, T10>,
+        InputSelector<TState, TProps, T11>,
+        InputSelector<TState, TProps, T12>,
+        InputSelector<TState, TProps, T13>,
+        InputSelector<TState, TProps, T14>,
+        InputSelector<TState, TProps, T15>,
+        InputSelector<TState, TProps, T16>
       ],
       resultFunc: (
         arg1: T1,
@@ -863,7 +869,7 @@ declare module "reselect" {
         arg15: T15,
         arg16: T16
       ) => TResult
-    ): Selector<TState, TProps, TResult>
+    ): OutputSelector<TState, TProps, TResult>
   };
 
   declare type Reselect = {
@@ -879,12 +885,10 @@ declare module "reselect" {
       ...memoizeOptions: any[]
     ) => SelectorCreator,
 
-    createStructuredSelector: <TState, TProps>(
-      inputSelectors: {
-        [k: string | number]: Selector<TState, TProps, any>
-      },
+    createStructuredSelector: <TState, TProps, InputSelectors: {[k: string | number]: InputSelector<TState, TProps, any>}>(
+      inputSelectors: InputSelectors,
       selectorCreator?: SelectorCreator
-    ) => Selector<TState, TProps, any>
+    ) => OutputSelector<TState, TProps, $ObjMap<InputSelectors, ExtractReturnType>>
   };
 
   declare module.exports: Reselect;
