@@ -1,5 +1,5 @@
-// flow-typed signature: 196c17cc5f7f821ccdc3dcd798742acf
-// flow-typed version: <<STUB>>/semantic-release_v^6.3.6/flow_v0.61.0
+// flow-typed signature: 7ebc2744c18a9bc12d5fdf3c0834d2d9
+// flow-typed version: <<STUB>>/semantic-release_v^15.1.2/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -26,35 +26,75 @@ declare module 'semantic-release/bin/semantic-release' {
   declare module.exports: any;
 }
 
-declare module 'semantic-release/src/lib/commits' {
+declare module 'semantic-release/cli' {
   declare module.exports: any;
 }
 
-declare module 'semantic-release/src/lib/get-registry' {
+declare module 'semantic-release/lib/definitions/errors' {
   declare module.exports: any;
 }
 
-declare module 'semantic-release/src/lib/plugin-noop' {
+declare module 'semantic-release/lib/definitions/plugins' {
   declare module.exports: any;
 }
 
-declare module 'semantic-release/src/lib/plugins' {
+declare module 'semantic-release/lib/definitions/release-types' {
   declare module.exports: any;
 }
 
-declare module 'semantic-release/src/lib/type' {
+declare module 'semantic-release/lib/get-commits' {
   declare module.exports: any;
 }
 
-declare module 'semantic-release/src/lib/verify' {
+declare module 'semantic-release/lib/get-config' {
   declare module.exports: any;
 }
 
-declare module 'semantic-release/src/post' {
+declare module 'semantic-release/lib/get-error' {
   declare module.exports: any;
 }
 
-declare module 'semantic-release/src/pre' {
+declare module 'semantic-release/lib/get-git-auth-url' {
+  declare module.exports: any;
+}
+
+declare module 'semantic-release/lib/get-last-release' {
+  declare module.exports: any;
+}
+
+declare module 'semantic-release/lib/get-next-version' {
+  declare module.exports: any;
+}
+
+declare module 'semantic-release/lib/git' {
+  declare module.exports: any;
+}
+
+declare module 'semantic-release/lib/hide-sensitive' {
+  declare module.exports: any;
+}
+
+declare module 'semantic-release/lib/logger' {
+  declare module.exports: any;
+}
+
+declare module 'semantic-release/lib/plugins/index' {
+  declare module.exports: any;
+}
+
+declare module 'semantic-release/lib/plugins/normalize' {
+  declare module.exports: any;
+}
+
+declare module 'semantic-release/lib/plugins/pipeline' {
+  declare module.exports: any;
+}
+
+declare module 'semantic-release/lib/utils' {
+  declare module.exports: any;
+}
+
+declare module 'semantic-release/lib/verify' {
   declare module.exports: any;
 }
 
@@ -62,27 +102,63 @@ declare module 'semantic-release/src/pre' {
 declare module 'semantic-release/bin/semantic-release.js' {
   declare module.exports: $Exports<'semantic-release/bin/semantic-release'>;
 }
-declare module 'semantic-release/src/lib/commits.js' {
-  declare module.exports: $Exports<'semantic-release/src/lib/commits'>;
+declare module 'semantic-release/cli.js' {
+  declare module.exports: $Exports<'semantic-release/cli'>;
 }
-declare module 'semantic-release/src/lib/get-registry.js' {
-  declare module.exports: $Exports<'semantic-release/src/lib/get-registry'>;
+declare module 'semantic-release/index' {
+  declare module.exports: $Exports<'semantic-release'>;
 }
-declare module 'semantic-release/src/lib/plugin-noop.js' {
-  declare module.exports: $Exports<'semantic-release/src/lib/plugin-noop'>;
+declare module 'semantic-release/index.js' {
+  declare module.exports: $Exports<'semantic-release'>;
 }
-declare module 'semantic-release/src/lib/plugins.js' {
-  declare module.exports: $Exports<'semantic-release/src/lib/plugins'>;
+declare module 'semantic-release/lib/definitions/errors.js' {
+  declare module.exports: $Exports<'semantic-release/lib/definitions/errors'>;
 }
-declare module 'semantic-release/src/lib/type.js' {
-  declare module.exports: $Exports<'semantic-release/src/lib/type'>;
+declare module 'semantic-release/lib/definitions/plugins.js' {
+  declare module.exports: $Exports<'semantic-release/lib/definitions/plugins'>;
 }
-declare module 'semantic-release/src/lib/verify.js' {
-  declare module.exports: $Exports<'semantic-release/src/lib/verify'>;
+declare module 'semantic-release/lib/definitions/release-types.js' {
+  declare module.exports: $Exports<'semantic-release/lib/definitions/release-types'>;
 }
-declare module 'semantic-release/src/post.js' {
-  declare module.exports: $Exports<'semantic-release/src/post'>;
+declare module 'semantic-release/lib/get-commits.js' {
+  declare module.exports: $Exports<'semantic-release/lib/get-commits'>;
 }
-declare module 'semantic-release/src/pre.js' {
-  declare module.exports: $Exports<'semantic-release/src/pre'>;
+declare module 'semantic-release/lib/get-config.js' {
+  declare module.exports: $Exports<'semantic-release/lib/get-config'>;
+}
+declare module 'semantic-release/lib/get-error.js' {
+  declare module.exports: $Exports<'semantic-release/lib/get-error'>;
+}
+declare module 'semantic-release/lib/get-git-auth-url.js' {
+  declare module.exports: $Exports<'semantic-release/lib/get-git-auth-url'>;
+}
+declare module 'semantic-release/lib/get-last-release.js' {
+  declare module.exports: $Exports<'semantic-release/lib/get-last-release'>;
+}
+declare module 'semantic-release/lib/get-next-version.js' {
+  declare module.exports: $Exports<'semantic-release/lib/get-next-version'>;
+}
+declare module 'semantic-release/lib/git.js' {
+  declare module.exports: $Exports<'semantic-release/lib/git'>;
+}
+declare module 'semantic-release/lib/hide-sensitive.js' {
+  declare module.exports: $Exports<'semantic-release/lib/hide-sensitive'>;
+}
+declare module 'semantic-release/lib/logger.js' {
+  declare module.exports: $Exports<'semantic-release/lib/logger'>;
+}
+declare module 'semantic-release/lib/plugins/index.js' {
+  declare module.exports: $Exports<'semantic-release/lib/plugins/index'>;
+}
+declare module 'semantic-release/lib/plugins/normalize.js' {
+  declare module.exports: $Exports<'semantic-release/lib/plugins/normalize'>;
+}
+declare module 'semantic-release/lib/plugins/pipeline.js' {
+  declare module.exports: $Exports<'semantic-release/lib/plugins/pipeline'>;
+}
+declare module 'semantic-release/lib/utils.js' {
+  declare module.exports: $Exports<'semantic-release/lib/utils'>;
+}
+declare module 'semantic-release/lib/verify.js' {
+  declare module.exports: $Exports<'semantic-release/lib/verify'>;
 }
