@@ -1,5 +1,5 @@
-// flow-typed signature: 3dab1daf5f4fac67aa2963034223c2d9
-// flow-typed version: <<STUB>>/shallowequal_v^1.0.2/flow_v0.61.0
+// flow-typed signature: c0148dd5b1eef4436f7bd3c751f7d926
+// flow-typed version: <<STUB>>/shallowequal_v^1.0.2/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
