@@ -1,5 +1,5 @@
-// flow-typed signature: 8844bc117c1823b690c64a5eeb486173
-// flow-typed version: <<STUB>>/style-loader_v^0.18.2/flow_v0.61.0
+// flow-typed signature: 63b04b8774beb6bd514267b3ec776bc0
+// flow-typed version: <<STUB>>/style-loader_v^0.18.2/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
