@@ -1,5 +1,5 @@
-// flow-typed signature: d6b3e9d6d59a07cdda47f9fe1d6e0f4d
-// flow-typed version: <<STUB>>/webpack_v^3.3.0/flow_v0.61.0
+// flow-typed signature: 80b272424a85bf16310e372a5db7669a
+// flow-typed version: <<STUB>>/webpack_v^3.3.0/flow_v0.85.0
 
 /**
  * This is an autogenerated libdef stub for:
