@@ -6,7 +6,7 @@ import { Router, Link, Route } from 'react-router-dom';
 import Root, { history } from '../shared/Root';
 import { formCreator } from '../Redux-Form/index';
 
-const Form1 = formCreator('form1');
+const Form1 = formCreator('form1', null);
 
 const FirstRoute = () => {
   return (
@@ -17,13 +17,13 @@ const FirstRoute = () => {
     </div>
   );
 };
-const Form2 = formCreator('form2');
+const Form2 = formCreator('form2', 'form2');
 
 const SecondRoute = () => {
   return (
     <div>
       <h1>Second-Route</h1>
-      <Link to="/first?form1.p.checked=true">Go to first</Link>
+      <Link to="/first?p.checked=true">Go to first</Link>
       <br />
       <Link to='/second?form2.p.checked=true&form2.p.text="initialText"'>Set some defaults</Link>
       <Form2 />
@@ -34,6 +34,7 @@ const SecondRoute = () => {
 const ReactRouter = () => {
   return (
     <Root>
+      { /* $FlowFixMe: history and router incompatible */}
       <Router history={history}>
         <div>
           <Link to="/first">Open initial</Link>
