@@ -4,29 +4,32 @@
 
 import React, { Component } from 'react';
 import type { ComponentType } from 'react';
-import PropTypes from 'prop-types';
+import { QueryManagerContext } from './components/QueryContainer';
+import type { QueryManager } from "./components/QueryContainer";
 
 const connectQuery =
   (makeRefAvailable: boolean = false) => (InnerComponent: ComponentType<*>) => {
     return class extends Component<*, *> {
-      static contextTypes = {
-        queryManager: PropTypes.object
-      };
 
       innerComponentRef: ?Element;
 
       render() {
-        const { queryManager } = this.context;
-        const additionalProps = makeRefAvailable ?
-          { ref: (instance) => { this.innerComponentRef = instance; } } : {};
         return (
-          <InnerComponent
-            createQueryString={queryManager.createQueryString}
-            persistCurrentState={queryManager.persistCurrentState}
-            __unsafe__queryManager={queryManager}
-            {...this.props}
-            {...additionalProps}
-          />
+          <QueryManagerContext.Consumer>
+            {(queryManager: ?QueryManager) => {
+              const additionalProps = makeRefAvailable ?
+              { ref: (instance) => { this.innerComponentRef = instance; } } : {};
+              return queryManager && (
+                <InnerComponent
+                  createQueryString={queryManager.createQueryString}
+                  persistCurrentState={queryManager.persistCurrentState}
+                  __unsafe__queryManager={queryManager}
+                  {...this.props}
+                  {...additionalProps}
+                />
+              );
+            }}
+          </QueryManagerContext.Consumer>
         );
       }
     };
